moved functions to separate namespace
