+++ plugins/bitorder_propagation/documentation/bitorder_propagation.rst
@@ -1,5 +1,8 @@
-Bitorder Propagation
+Bit-Order Propagation
 ==========================
+
+.. automodule:: bitorder_propagation
+    :members: reorder_module_pin_groups, propagate_bitorder
 
 .. autoclass:: bitorder_propagation.BitorderPropagationPlugin
    :members:
